json 오류 수정

+++ src/main/java/com/jangayo/study/cart/form/OptionForm.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class OptionForm {
 
     private Long optionId;
-    private Long cnt;
+    private int cnt;
     
 }
